adding some more tests and increasing code coverage

+++ src/driver/tests/mod.rs
@@ -59,19 +59,28 @@ fn fixture_test(test_name: &str) {
 
     if std::env::var("OVERWRITE_FIXTURES").is_ok() {
         for (actual_path, expected_path) in expected_actual_pairs.iter() {
-            let actual = std::fs::read_to_string(&actual_path).unwrap();
+            if let Ok(actual) = std::fs::read_to_string(&actual_path) {
 
-            if let Ok(mut expected_file) = File::create(&expected_path) {
-                expected_file.write_all(actual.as_bytes()).unwrap();
+                if let Ok(mut expected_file) = File::create(&expected_path) {
+                    expected_file.write_all(actual.as_bytes()).unwrap();
+                }
+            } else {
+                File::create(&expected_path).unwrap();
             }
         }
     }
 
     for (actual_path, expected_path) in expected_actual_pairs.iter() {
-        let actual = std::fs::read_to_string(&actual_path).unwrap();
-        let expected = std::fs::read_to_string(&expected_path).unwrap();
+        if let Ok(actual) = std::fs::read_to_string(&actual_path) {
+            let expected = std::fs::read_to_string(&expected_path).unwrap();
+            //let actual = std::fs::read_to_string(&actual_path).unwrap();
 
-        assert_eq!(expected, actual);
-        std::fs::remove_file(&actual_path).unwrap();
+            assert_eq!(expected, actual);
+            std::fs::remove_file(&actual_path).unwrap();
+        } else {
+            let expected = std::fs::read_to_string(&expected_path).unwrap();
+
+            assert_eq!(expected, "");
+        }
     }
 }
